commit: checkpoint 17/2/2025 1:44 am

+++ Assets/com.ethnicthv/chemlab/engine/api/mixture/IMixture.cs
@@ -6,11 +6,11 @@ namespace com.ethnicthv.chemlab.engine.api.mixture
     {
         public MixtureType GetMixtureType();
         public void Tick();
-        public void AddMolecule(Molecule molecule, double moles);
+        public void AddMolecule(Molecule molecule, float moles);
         public void RemoveMolecule(Molecule molecule);
-        public void SetMoles(Molecule molecule, double moles);
-        public double GetMoles(Molecule molecule);
-        public double AddMoles(Molecule molecule, double moles, out bool isMutating);
-        public double SubtractMoles(Molecule molecule, double moles, out bool isMutating);
+        public void SetMoles(Molecule molecule, float moles);
+        public float GetMoles(Molecule molecule);
+        public float AddMoles(Molecule molecule, float moles, out bool isMutating);
+        public float SubtractMoles(Molecule molecule, float moles, out bool isMutating);
     }
 }
